Implement float_to_list/2 accepting a list of formatting options
The function extends float_to_list/1 by adding a second Options
argument. Returns a string which corresponds to the text
representation of Float using fixed decimal point formatting
rounded to the Decimals number of digits.

    float_to_list(Float, Options) -> string()

        Float   = float()
        Options = [Option]
        Option  = {decimals, Decimals::integer()} | compact

The 'decimals' option controls the number of digits after the decimal
point in the output.  The 'compact' option truncates the trailing zeros
at the end of the returned string.

This BIF is optimized to run 5-10x faster than float_to_list/1 for
floats under 2^52.  Formatting of floats larger than that value is
done using legacy implementation.