Fix 2 second delay after eval in shell running under ssh_cli
ssh_cli needs to reply to the tty_geometry message or else group
will wait 2000 msec every time an expression is successfully
evaluated in the shell.

A "more correct" fix might refactor io_request and io_requests to
take Group as an argument so that it isn't special-cased in
handle_msg.